+++ src/index.js
@@ -3,4 +3,5 @@ export { default as fetchState } from './components/fetchState';
 export { default as Module } from './components/Module';
 export { default as preload } from './module/preload';
 export { default as renderToString } from './renderer/renderToString';
+export { default as renderToStaticMarkup } from './renderer/renderToStaticMarkup';
 export { default as ServerStateProvider } from './components/ServerStateProvider';